+++ src/PhpWord/Writer/Word2007/Footnotes.php
@@ -29,15 +29,8 @@ class Footnotes extends Base
     public function writeFootnotes($allFootnotesCollection)
     {
         // Create XML writer
-        $xmlWriter = null;
-        if ($this->getParentWriter()->getUseDiskCaching()) {
-            $xmlWriter = new XMLWriter(
-                XMLWriter::STORAGE_DISK,
-                $this->getParentWriter()->getDiskCachingDirectory()
-            );
-        } else {
-            $xmlWriter = new XMLWriter(XMLWriter::STORAGE_MEMORY);
-        }
+        $xmlWriter = $this->getXmlWriter();
+
         // XML header
         $xmlWriter->startDocument('1.0', 'UTF-8', 'yes');
         $xmlWriter->startElement('w:footnotes');