+++ src/PhpWord/Writer/Word2007/Styles.php
@@ -34,17 +34,11 @@ class Styles extends Base
      */
     public function writeStyles(PhpWord $phpWord = null)
     {
-        // Create XML writer
-        $xmlWriter = null;
-        if ($this->getParentWriter()->getUseDiskCaching()) {
-            $xmlWriter = new XMLWriter(
-                XMLWriter::STORAGE_DISK,
-                $this->getParentWriter()->getDiskCachingDirectory()
-            );
-        } else {
-            $xmlWriter = new XMLWriter(XMLWriter::STORAGE_MEMORY);
-        }
         $this->phpWord = $phpWord;
+
+        // Create XML writer
+        $xmlWriter = $this->getXmlWriter();
+
         // XML header
         $xmlWriter->startDocument('1.0', 'UTF-8', 'yes');
         $xmlWriter->startElement('w:styles');