+++ CHANGELOG.md
@@ -21,6 +21,8 @@ This is the changelog between releases of PHPWord. Releases are listed in revers
 ### Miscellaneous
 
 - Documentation: Simplify page level docblock - @ivanlanin GH-179
+- Writer: Refactor writer classes and make a new Writer abstract class - @ivanlanin GH-160
+- Reader: Rename AbstractReader > Reader - @ivanlanin
 
 ## 0.9.1 - 27 Mar 2014